refactor: rename package from server to websocket
