+++ start/src/main/java/org/jakartaee/sample/websocket/event/GameServerEvent.java
@@ -1,4 +1,4 @@
-package org.jakartaee.sample.server.event;
+package org.jakartaee.sample.websocket.event;
 
 public sealed interface GameServerEvent
         permits OnSessionOpenEvent, OnSessionMessageEvent, OnSessionCloseEvent, OnSessionErrorEvent {
