+++ start/src/main/java/org/jakartaee/sample/websocket/listener/OnSessionMessageEventListener.java
@@ -1,11 +1,11 @@
-package org.jakartaee.sample.server.listener;
+package org.jakartaee.sample.websocket.listener;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
-import org.jakartaee.sample.server.GameSessions;
-import org.jakartaee.sample.server.event.OnSessionMessageEvent;
-import org.jakartaee.sample.server.message.Message;
+import org.jakartaee.sample.websocket.GameSessions;
+import org.jakartaee.sample.websocket.event.OnSessionMessageEvent;
+import org.jakartaee.sample.websocket.message.Message;
 
 import static java.util.Objects.isNull;
