+++ start/src/main/java/org/jakartaee/sample/websocket/listener/OnSessionOpenEventListener.java
@@ -1,10 +1,10 @@
-package org.jakartaee.sample.server.listener;
+package org.jakartaee.sample.websocket.listener;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
-import org.jakartaee.sample.server.GameSessions;
-import org.jakartaee.sample.server.event.OnSessionOpenEvent;
+import org.jakartaee.sample.websocket.GameSessions;
+import org.jakartaee.sample.websocket.event.OnSessionOpenEvent;
 
 @ApplicationScoped
 public class OnSessionOpenEventListener {
