+++ start/src/main/java/org/jakartaee/sample/websocket/message/MessageField.java
@@ -1,4 +1,4 @@
-package org.jakartaee.sample.server.message;
+package org.jakartaee.sample.websocket.message;
 
 import java.util.Optional;
